- Remove ToSuperScript for double as we can't actually represent double values as there are no superscript chars for period and characters - Add comments to the Extensions methods - Add more tests for ToSuperScript

+++ Measurement/Quantity.cs
@@ -703,7 +703,7 @@ namespace ForgedSoftware.Measurement {
 				if (eIndex >= 0) {
 					double exponent = Math.Floor(Math.Log(Value) / Math.Log(10));
 					valueStr = valueStr.Substring(0, eIndex);
-					string exponentStr = (options.Ascii) ? "^" + exponent : exponent.ToSuperScript();
+					string exponentStr = (options.Ascii) ? "^" + exponent : ((int)exponent).ToSuperScript();
 					valueStr += " x 10" + exponentStr;
 				}
 			}
